+++ src/components/componentGroupButton/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-group-button
 
+## 0.0.3
+
+### Patch Changes
+
+- Setting a property to change emojis styles
+
 ## 0.0.2
 
 ### Patch Changes