+++ src/components/pageBroteApp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # brote-app
 
+## 0.0.16
+
+### Patch Changes
+
+- Setting a property to change emojis styles
+
 ## 0.0.15
 
 ### Patch Changes